+++ probes/rust-vertical-slice/consumer/src/lib.rs
@@ -1,4 +1,0 @@
-#![allow(dead_code)]
-
-#[path = "../../../../contracts/product/v1/rust-sdk-consumer.rs"]
-mod public_sdk_contract;